+++ lib/dragonfly_pdf/plugin.rb
@@ -1,5 +1,5 @@
 require 'dragonfly_pdf/analysers/pdf_properties'
-require 'dragonfly_pdf/processors/combine'
+require 'dragonfly_pdf/processors/append'
 require 'dragonfly_pdf/processors/page'
 require 'dragonfly_pdf/processors/page_thumb'
 require 'dragonfly_pdf/processors/stamp'
@@ -36,7 +36,7 @@ module DragonflyPdf
 
       # ---------------------------------------------------------------------
 
-      app.add_processor :combine, DragonflyPdf::Processors::Combine.new
+      app.add_processor :append, DragonflyPdf::Processors::Append.new
       app.add_processor :page, DragonflyPdf::Processors::Page.new
       app.add_processor :page_thumb, DragonflyPdf::Processors::PageThumb.new
       app.add_processor :stamp, DragonflyPdf::Processors::Stamp.new