+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.1
 
-* added: `combine` which allows to combine multiplePDFs together
+* added: `append` which allows to combine multiple PDFs together
 * added: `stamp` which allows for stamping all pages of a PDF with content of another PDF
 
 ## 0.1.0